+++ Engine/Poo.cpp
@@ -2,7 +2,7 @@
 #include "Graphics.h"
 #include <assert.h>
 
-void Poo::Init( int in_x,int in_y,int in_vx,int in_vy )
+void Poo::Init( float in_x,float in_y,float in_vx,float in_vy )
 {
 	assert( initialized == false );
 	x = in_x;
@@ -18,7 +18,7 @@ void Poo::Update()
 	x += vx;
 	y += vy;
 
-	const int right = x + width;
+	const float right = x + width;
 	if( x < 0 )
 	{
 		x = 0;
@@ -30,7 +30,7 @@ void Poo::Update()
 		vx = -vx;
 	}
 
-	const int bottom = y + height;
+	const float bottom = y + height;
 	if( y < 0 )
 	{
 		y = 0;
@@ -48,8 +48,8 @@ bool Poo::TestCollision( const Dude& dude ) const
 	assert( initialized == true );
 	const int duderight = dude.GetX() + dude.GetWidth();
 	const int dudebottom = dude.GetY() + dude.GetHeight();
-	const int pooright = x + width;
-	const int poobottom = y + height;
+	const float pooright = x + width;
+	const float poobottom = y + height;
 
 	return
 		duderight >= x &&
